Update main.sh
The error message indicates that the repository git@github.com:hellman/hacking-tools.git does not exist. This could be due to a typo or an incorrect repository URL. Let's correct this by using a valid repository URL for hacking tools.

I'll update the script to use a different, valid repository for hacking tools. We'll use the HackingTools repository from carlospolop as an example: